:hammer: code refactoring on data loader
